Update md5tool.py
Trying to add a function to create one master file that lists al the md5 hashes with the filenames and file paths in a directory. Intention is to have this md5 master file as a tool to compare file differences in two directory's that are on two different computers, without computer A having to read computer B. I am still learning, I am new to Python as well as GitHub.